Support globs and directories in `extra_files`
All entries in `extra_files` are now interpreted as zglob-style globs. Additionally, directories can
be specified and will be expanded to a recursive glob under the directory. File path handling is the
same as it has always been.

Hidden files are not included in globs and must be spelled out explicitly.